test: Only run rexpect tests on unix
